pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "ibis-framework[pandas,polars]>=9.5.0",
    "ipykernel>=6.29.5",
    "streamlit>=1.40.2",
]

[tool.ruff]
select =["ALL"]
